implement caching, and some other major improvements
With SoC, another separational layer was implemented: Containers. These classes can hide the implementation (e.g. for the user it doesn't matter, whether SDWebImage or Nuke downloads the data, only that it must work). Currently there are 3 ServiceContainers (API (HTTP) communication, persistence and imagedownloaders). With this concept in mind it can be easily broaden with some other libraries or custom implementations.